requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
Pillow
numpy
python-multipart
easyocr
torch
tesseract
